requirements.txt
Django>=6.0,<6.1
python-dotenv>=1.0.1
groq>=0.11.0
djangorestframework>=3.16.0
djangorestframework-simplejwt>=5.4.0
django-cors-headers>=4.6.0
firebase-admin>=6.6.0
Pillow>=10.4.0
pytesseract>=0.3.13
